Update Lab 4 Exercice 2 - IRM Validation
There was a misunderstanding in the IRM principle : IRM, as setup on the SharePoint library in this exercice, will protect downloaded copies of documents to align to the user's SharePoint permissions. (the user not being able to edit a protected document with Word Online is just a side-effect : IRM protected documents can't be edited in Office Online).
I propose a change in the exercice setup and test to show the real IRM behavior.